Split up WS2812FX.cpp in FX and helper files
