+++ wled00/wled00.ino
@@ -78,7 +78,7 @@
 
 
 //version code in format yymmddb (b = daily build)
-#define VERSION 1902091
+#define VERSION 1902092
 char versionString[] = "0.8.3-dev";
